update tleap class
