+++ docs/source/tleap.rst
@@ -1,8 +1,17 @@
-TLeap API
----------
+System API
+==========
 
-.. automodule:: paprika.build.tleap
+TLeap Wrapper
+-------------
+
+.. automodule:: paprika.build.system.tleap
    :members:
    :undoc-members:
    :show-inheritance:
 
+Utils
+-----
+.. automodule:: paprika.build.system.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
