+++ paprika/tests/test_tleap.py
@@ -12,8 +12,8 @@ import parmed as pmd
 import pytest
 
 from paprika.build.align import zalign
-from paprika.build.dummy import add_dummy, write_dummy_frcmod, write_dummy_mol2
-from paprika.build.tleap import ANGSTROM_CUBED_TO_LITERS, System
+from paprika.build.system import TLeap
+from paprika.build.system.utils import ANGSTROM_CUBED_TO_LITERS, PBCBox
 from paprika.utils import is_file_and_not_empty
 
 logger = logging.getLogger(__name__)
@@ -27,7 +27,7 @@ def clean_files(directory=os.path.join(os.path.dirname(__file__), "tmp")):
     os.makedirs(directory)
     yield
     # This happens after the test function call
-    shutil.rmtree(directory)
+    # shutil.rmtree(directory)
 
 
 @pytest.mark.slow
@@ -35,7 +35,7 @@ def test_solvation_simple(clean_files):
     """ Test that we can solvate CB6-BUT using default settings. """
     waters = np.random.randint(100, 10000)
     logger.debug("Trying {} waters with default settings...".format(waters))
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -49,12 +49,12 @@ def test_solvation_simple(clean_files):
     assert int(grepped_waters) == waters
 
 
-@pytest.mark.parametrize("shape", ["octahedral", "cubic"])
+@pytest.mark.parametrize("shape", [PBCBox.octahedral, PBCBox.cubic])
 def test_solvation_shapes(shape, clean_files):
     """ Test that we can solvate CB6-BUT with a truncated octahedron. """
     waters = np.random.randint(1000, 10000)
     logger.debug("Trying {} waters in a truncated octahedron...".format(waters))
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -72,13 +72,42 @@ def test_solvation_shapes(shape, clean_files):
     assert int(grepped_waters) == waters
 
 
+@pytest.mark.parametrize("water_model", ["tip4p", "opc"])
+def test_solvation_water_model(water_model, clean_files):
+    """ Test that we can solvate CB6-BUT with a truncated octahedron. """
+    waters = np.random.randint(1000, 10000)
+    logger.debug("Trying {} waters in a truncated octahedron...".format(waters))
+    sys = TLeap()
+    sys.output_path = "tmp"
+    sys.loadpdb_file = os.path.join(
+        os.path.dirname(__file__), "../data/cb6-but/cb6-but.pdb"
+    )
+    sys.target_waters = waters
+    sys.output_prefix = "solvate"
+    sys.pbc_type = PBCBox.cubic
+    sys.set_water_model(water_model)
+    sys.template_lines = [
+        "source leaprc.protein.ff14SB",
+        "loadamberparams ../../data/cb6-but/cb6.frcmod",
+        "CB6 = loadmol2 ../../data/cb6-but/cb6.mol2",
+        "loadamberparams ../../data/cb6-but/but.frcmod",
+        "BUT = loadmol2 ../../data/cb6-but/but.mol2",
+        "model = loadpdb ../../data/cb6-but/cb6-but.pdb",
+    ]
+    sys.build()
+    grepped_waters = sp.check_output(
+        ["grep -oh 'WAT' ./tmp/solvate.prmtop | wc -w"], shell=True
+    )
+    assert int(grepped_waters) == waters
+
+
 @pytest.mark.slow
 def test_solvation_spatial_size(clean_files):
     """ Test that we can solvate CB6-BUT with an buffer size in Angstroms. """
     random_int = np.random.randint(10, 20)
     random_size = random_int * np.random.random_sample(1) + random_int
     logger.debug("Trying buffer size of {} A...".format(random_size[0]))
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -88,7 +117,7 @@ def test_solvation_spatial_size(clean_files):
     )
     sys.buffer_value = float(random_size[0])
     sys.output_prefix = "solvate"
-    sys.pbc_type = "cubic"
+    sys.pbc_type = PBCBox.cubic
     sys.build()
     grepped_waters = sp.check_output(
         ["grep -oh 'WAT' ./tmp/solvate.prmtop | wc -w"], shell=True
@@ -101,7 +130,7 @@ def test_solvation_potassium_control(clean_files):
     """ Test there is no potassium by default. A negative control. """
     waters = np.random.randint(1000, 10000)
     logger.debug("Trying {} waters with potassium...".format(waters))
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -130,7 +159,7 @@ def test_solvation_with_additional_ions(clean_files):
     random_cation = random.choice(cations)
     random_anion = random.choice(anions)
     logger.debug("Trying {} waters with additional ions...".format(waters))
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -171,7 +200,7 @@ def test_solvation_with_additional_ions(clean_files):
 def test_solvation_by_M_and_m(clean_files):
     """ Test that we can solvate CB6-BUT through molarity and molality. """
     logger.debug("Trying 10 A buffer with 150 mM NaCl...")
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -182,7 +211,7 @@ def test_solvation_by_M_and_m(clean_files):
     sys.buffer_value = 10.0
     sys.output_prefix = "solvate"
     sys.neutralize = False
-    sys.pbc_type = "rectangular"
+    sys.pbc_type = PBCBox.rectangular
     sys.add_ions = ["NA", "0.150M", "CL", "0.150M", "K", "0.100m", "BR", "0.100m"]
     sys.build()
 
@@ -198,8 +227,8 @@ def test_solvation_by_M_and_m(clean_files):
 
     volume = sys.get_volume()
     volume_in_liters = volume * ANGSTROM_CUBED_TO_LITERS
-    calc_num_na = np.ceil((6.022 * 10 ** 23) * (0.150) * volume_in_liters)
-    calc_num_cl = np.ceil((6.022 * 10 ** 23) * (0.150) * volume_in_liters)
+    calc_num_na = np.ceil((6.022 * 10 ** 23) * 0.150 * volume_in_liters)
+    calc_num_cl = np.ceil((6.022 * 10 ** 23) * 0.150 * volume_in_liters)
     assert int(obs_num_na) == int(calc_num_na)
     assert int(obs_num_cl) == int(calc_num_cl)
 
@@ -229,7 +258,7 @@ def test_alignment_workflow(clean_files):
     )
     zalign(cb6, ":CB6", ":BUT", save=True, filename="./tmp/tmp.pdb")
     waters = np.random.randint(1000, 10000)
-    sys = System()
+    sys = TLeap()
     sys.template_file = os.path.join(
         os.path.dirname(__file__), "../data/cb6-but/tleap_solvate.in"
     )
@@ -245,80 +274,10 @@ def test_alignment_workflow(clean_files):
     assert int(grepped_waters) == waters
 
 
-def test_add_dummy(clean_files):
-    """ Test that dummy atoms get added correctly """
-    temporary_directory = os.path.join(os.path.dirname(__file__), "tmp")
-    host_guest = pmd.load_file(
-        os.path.join(
-            os.path.dirname(__file__), "../data/cb6-but/cb6-but-notcentered.pdb"
-        ),
-        structure=True,
-    )
-    host_guest = zalign(host_guest, ":BUT@C", ":BUT@C3", save=False)
-    host_guest = add_dummy(host_guest, residue_name="DM1", z=-11.000, y=2.000, x=-1.500)
-
-    host_guest.write_pdb(
-        os.path.join(temporary_directory, "cb6-but-dum.pdb"), renumber=False
-    )
-    with open(os.path.join(temporary_directory, "cb6-but-dum.pdb"), "r") as f:
-        lines = f.readlines()
-        test_line1 = lines[123].rstrip()
-        test_line2 = lines[124].rstrip()
-    ref_line1 = "TER     123      BUT     2"
-    ref_line2 = (
-        "HETATM  123 DUM  DM1     3      -1.500   2.000 -11.000  0.00  0.00          PB"
-    )
-    assert ref_line1 == test_line1
-    assert ref_line2 == test_line2
-
-    write_dummy_frcmod(path=temporary_directory)
-    write_dummy_mol2(path=temporary_directory, filename="dm1.mol2", residue_name="DM1")
-    sys = System()
-    cb6_frcmod = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "../data/cb6-but/cb6.frcmod")
-    )
-    cb6_mol2 = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "../data/cb6-but/cb6.mol2")
-    )
-    but_frcmod = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "../data/cb6-but/but.frcmod")
-    )
-    but_mol2 = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "../data/cb6-but/but.mol2")
-    )
-
-    sys.template_lines = [
-        "source leaprc.gaff",
-        f"loadamberparams {cb6_frcmod}",
-        f"CB6 = loadmol2 {cb6_mol2}",
-        f"loadamberparams {but_frcmod}",
-        f"BUT = loadmol2 {but_mol2}",
-        "loadamberparams dummy.frcmod",
-        "DM1 = loadmol2 dm1.mol2",
-        "model = loadpdb cb6-but-dum.pdb",
-    ]
-    sys.output_path = temporary_directory
-    sys.output_prefix = "cb6-but-dum"
-    sys.pbc_type = None
-    sys.neutralize = False
-    sys.build()
-    with open(
-        os.path.join(os.path.dirname(__file__), "../data/cb6-but/REF_cb6-but-dum.rst7"),
-        "r",
-    ) as f:
-        contents = f.read()
-        reference = [float(i) for i in contents.split()[2:]]
-
-    with open(os.path.join(temporary_directory, "cb6-but-dum.rst7"), "r") as f:
-        contents = f.read()
-        new = [float(i) for i in contents.split()[2:]]
-    assert np.allclose(reference, new)
-
-
 def test_hydrogen_mass_repartitioning(clean_files):
     """ Test that hydrogen mass is repartitioned. """
     temporary_directory = os.path.join(os.path.dirname(__file__), "tmp")
-    sys = System()
+    sys = TLeap()
     but_frcmod = os.path.abspath(
         os.path.join(os.path.dirname(__file__), "../data/cb6-but/but.frcmod")
     )
@@ -356,7 +315,7 @@ def test_multiple_pdb_files(clean_files):
     Reference: https://github.com/slochower/pAPRika/issues/141
     """
     temporary_directory = os.path.join(os.path.dirname(__file__), "tmp")
-    sys = System()
+    sys = TLeap()
     cb6_frcmod = os.path.abspath(
         os.path.join(os.path.dirname(__file__), "../data/cb6-but/cb6.frcmod")
     )
@@ -406,7 +365,7 @@ def test_conversions(clean_files):
         os.path.join(os.path.dirname(__file__), "../data/cb6-but/but.mol2")
     )
 
-    sys = System()
+    sys = TLeap()
     sys.template_lines = [
         "source leaprc.gaff",
         f"loadamberparams {but_frcmod}",
@@ -419,7 +378,7 @@ def test_conversions(clean_files):
     sys.neutralize = False
     sys.build()
 
-    from paprika.build.tleap import ConversionToolkit
+    from paprika.build.system.utils import ConversionToolkit
     from paprika.utils import is_file_and_not_empty
 
     # Gromacs ParmEd test
@@ -505,10 +464,10 @@ def test_ion_solvation(clean_files):
             )
 
         # Run tleap
-        ion_sys = System()
+        ion_sys = TLeap()
         ion_sys.output_path = temporary_directory
         ion_sys.target_waters = 1000
-        ion_sys.pbc_type = "cubic"
+        ion_sys.pbc_type = PBCBox.cubic
         ion_sys.neutralize = False
         ion_sys.template_lines = [
             "source leaprc.gaff",
