+++ paprika/tests/tmp/build.tleap.in
@@ -1,0 +1,16 @@
+source leaprc.water.tip3p
+loadamberparams frcmod.tip3p
+source leaprc.gaff
+source leaprc.water.tip3p
+Cl- = loadmol2 ion.mol2
+model = loadpdb ion.pdb
+
+solvatebox model TIP3PBOX 14.250000000000002 iso
+remove model model.1002
+remove model model.1003
+remove model model.1004
+remove model model.1005
+savepdb model build.pdb
+saveamberparm model build.prmtop build.rst7
+desc model
+quit
