+++ paprika/tests/tmp/ion.pdb
@@ -1,0 +1,3 @@
+CRYST1   30.000   30.000   30.000  90.00  90.00  90.00 P 1           1
+ATOM      1 Cl-  Cl-     1       0.000   0.000   0.000  1.00  0.00
+END
